Created ExternalMemoryTreap data structure.
It consists of a set of regular Treaps where each one has it's own
non-overlapping key range. When a treap gets too large, it is split in
half and it's split point is saved to determine the new key range.
These sub-treaps are cached out to disk when a specified memory limit
is reached.